将抽象语法树节点中的 enum 枚举类移动到 static 中
